Split up pocl_llvm_api.cc into multiple files
File has grown too big, and compile times too. Split up into
four files: utils, metadata-, workgroup(parallel.bc)-related, and
build/link(program.bc)-related tasks.

Added pocl_llvm_api.h to connect them all (unlike pocl_llvm.h it
exports C++ code, not C).
