+++ lib/CL/CMakeLists.txt
@@ -154,8 +154,9 @@ add_library("pocl_cache" OBJECT "pocl_cache.c")
 
 if (OCS_AVAILABLE)
   include_directories(${LLVM_INCLUDE_DIRS})
-  list(APPEND POCL_LIB_SOURCES "pocl_llvm_api.cc")
-  set_source_files_properties("pocl_llvm_api.cc" PROPERTIES COMPILE_FLAGS "${LLVM_CXXFLAGS} -I\"${CMAKE_CURRENT_SOURCE_DIR}/../llvmopencl\"")
+  set(LLVM_API_SOURCES "pocl_llvm_build.cc" "pocl_llvm_metadata.cc" "pocl_llvm_utils.cc" "pocl_llvm_wg.cc")
+  list(APPEND POCL_LIB_SOURCES ${LLVM_API_SOURCES})
+  set_source_files_properties(${LLVM_API_SOURCES} PROPERTIES COMPILE_FLAGS "${LLVM_CXXFLAGS} -I\"${CMAKE_CURRENT_SOURCE_DIR}/../llvmopencl\"")
   list(APPEND LIBPOCL_OBJS "$<TARGET_OBJECTS:llvmpasses>")
   # pocl_cache.c depends on a SHA1 hash of all built kernel-<target>.bc
   add_dependencies("pocl_cache" "kernellib_hash")
