+++ lib/CL/pocl_llvm_api.h
@@ -1,0 +1,58 @@
+/* pocl_llvm_api.cc: internally used header for pocl's LLVM API sources.
+
+   Copyright (c) 2013 Kalle Raiskila
+                 2013-2017 Pekka Jääskeläinen
+
+   Permission is hereby granted, free of charge, to any person obtaining a copy
+   of this software and associated documentation files (the "Software"), to deal
+   in the Software without restriction, including without limitation the rights
+   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+   copies of the Software, and to permit persons to whom the Software is
+   furnished to do so, subject to the following conditions:
+
+   The above copyright notice and this permission notice shall be included in
+   all copies or substantial portions of the Software.
+
+   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+   THE SOFTWARE.
+*/
+
+#include "pocl_llvm.h"
+
+#include <llvm/Support/Mutex.h>
+#include <llvm/IR/Module.h>
+#include <map>
+#include <string>
+
+#ifdef __GNUC__
+#pragma GCC visibility push(hidden)
+#endif
+
+extern llvm::sys::Mutex kernelCompilerLock;
+
+typedef struct _cl_device_id *cl_device_id;
+extern cl_device_id currentPoclDevice;
+
+void InitializeLLVM();
+llvm::LLVMContext &GlobalContext();
+extern long numberOfIRs;
+
+llvm::Module *parseModuleIR(const char *path);
+void writeModuleIR(const llvm::Module *mod, std::string &str);
+llvm::Module *parseModuleIRMem(const char *input_stream, size_t size);
+int getModuleTriple(const char *input_stream, size_t size, std::string &triple);
+
+void clearKernelPasses();
+void clearTargetMachines();
+void cleanKernelLibrary();
+
+extern std::string currentWgMethod;
+
+#ifdef __GNUC__
+#pragma GCC visibility pop
+#endif
